requirements.txt
Flask
flask-sqlalchemy
sqlalchemy
Flask-Login
plotly
pandas
numpy
python-dateutil
natsort
Flask-Migrate
waitress
werkzeug
passkeys
python-dotenv
cryptography
openai